chore: improve folder and file structures
