+++ packages/sdk/cloudflare/jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   transform: { '^.+\\.ts?$': 'ts-jest' },
   testMatch: ['**/*.test.ts?(x)'],
-  testPathIgnorePatterns: ['node_modules', 'example'],
+  testPathIgnorePatterns: ['node_modules', 'example', 'dist'],
   testEnvironment: 'node',
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   collectCoverageFrom: ['src/**/*.ts'],
