+++ packages/sdk/cloudflare/src/createLDClient/createCallbacks.ts
@@ -1,0 +1,19 @@
+import { EventEmitter } from 'node:events';
+
+const createCallbacks = (emitter: EventEmitter) => ({
+  onError: (err: Error) => {
+    if (emitter.listenerCount('error')) {
+      emitter.emit('error', err);
+    }
+  },
+  onFailed: (err: Error) => {
+    emitter.emit('failed', err);
+  },
+  onReady: () => {
+    emitter.emit('ready');
+  },
+  onUpdate: () => {},
+  hasEventListeners: () => false,
+});
+
+export default createCallbacks;
